// src/openapi/model/api-orderer-project-application.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import {ApiContractor} from './api-contractor';
import {ApiOrdererProject} from './api-orderer-project';

/**
 *
 * @export
 * @interface ApiOrdererProjectApplication
 */
export interface ApiOrdererProjectApplication {
    /**
     *
     * @type {string}
     * @memberof ApiOrdererProjectApplication
     */
    recruitingStartDate: string;
    /**
     *
     * @type {string}
     * @memberof ApiOrdererProjectApplication
     */
    recruitingEndDate: string;
    /**
     *
     * @type {string}
     * @memberof ApiOrdererProjectApplication
     */
    projectName: string;
    /**
     *
     * @type {number}
     * @memberof ApiOrdererProjectApplication
     */
    ordererPrice: number;
    /**
     *
     * @type {number}
     * @memberof ApiOrdererProjectApplication
     */
    contractorId: number;
    /**
     *
     * @type {number}
     * @memberof ApiOrdererProjectApplication
     */
    projectApplicationStatus: number;
    /**
     *
     * @type {string}
     * @memberof ApiOrdererProjectApplication
     */
    workEndDate: string;
    /**
     *
     * @type {ApiOrdererProject}
     * @memberof ApiOrdererProjectApplication
     */
    project?: ApiOrdererProject;
    /**
     *
     * @type {ApiContractor}
     * @memberof ApiOrdererProjectApplication
     */
    contractor?: ApiContractor;
}